refactoring functions
